requirements.txt
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
Requests==2.31.0
streamlit==1.33.0
ultralytics==8.2.2
